+++ custom_model_runner/datarobot_drum/drum/description.py
@@ -1,3 +1,3 @@
-version = "1.3.0"
+version = "1.3.1rc2"
 __version__ = version
 project_name = "datarobot-drum"