+++ custom_model_runner/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.3.1rc1] - in progress
+##### Added
+- `multiclass` target type
+- `class-labels` for specifying the class labels of a multiclass model
+- `class-labels-file` for specifying a file containing the class labels of a multiclass model
+##### Changed
+- support only keras built into tensorflow >= 2.2.1
 
 #### Unreleased
 ##### Added